practise1/requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
certifi==2026.1.4
charset-normalizer==3.4.4
contourpy==1.3.2
cycler==0.12.1
filelock==3.20.2
fonttools==4.61.1
fsspec==2025.12.0
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==12.1.0
polars==1.36.1
polars-runtime-32==1.36.1
psutil==7.2.1
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scipy==1.15.3
simsimd==6.5.12
six==1.17.0
stringzilla==4.6.0
sympy==1.14.0
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.3.249
ultralytics-thop==2.0.18
urllib3==2.6.2
albucore==0.0.24
albumentations==2.0.8
alembic==1.18.1
annotated-types==0.7.0
certifi==2026.1.4
charset-normalizer==3.4.4
contourpy==1.3.2
cycler==0.12.1
filelock==3.20.2
fonttools==4.61.1
fsspec==2025.12.0
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==12.1.0
polars==1.36.1
polars-runtime-32==1.36.1
psutil==7.2.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.1
PySide6==6.10.1
PySide6_Addons==6.10.1
PySide6_Essentials==6.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scipy==1.15.3
shiboken6==6.10.1
simsimd==6.5.12
six==1.17.0
SQLAlchemy==2.0.45
stringzilla==4.6.0
sympy==1.14.0
tomli==2.4.0
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.3.249
ultralytics-thop==2.0.18
urllib3==2.6.2
